Streamfield updates
* added blocks that map 1:1 to old static articles
* revamped styling of all blocks to fit in the template
* refactor story streamfield
* extract streamfield fields into a separate module
