+++ blacktail/models/fields.py
@@ -1,0 +1,78 @@
+from django.db import models
+from wagtail.admin.edit_handlers import FieldPanel, MultiFieldPanel, PageChooserPanel
+from wagtail.documents.edit_handlers import DocumentChooserPanel
+
+
+class ContactFields(models.Model):
+    telephone = models.CharField(max_length=20, blank=True)
+    email = models.EmailField(blank=True)
+    address_1 = models.CharField(max_length=255, blank=True)
+    address_2 = models.CharField(max_length=255, blank=True)
+    city = models.CharField(max_length=255, blank=True)
+    country = models.CharField(max_length=255, blank=True)
+    post_code = models.CharField(max_length=10, blank=True)
+
+    panels = [
+        FieldPanel('telephone'),
+        FieldPanel('email'),
+        FieldPanel('address_1'),
+        FieldPanel('address_2'),
+        FieldPanel('city'),
+        FieldPanel('country'),
+        FieldPanel('post_code'),
+    ]
+
+    api_fields = ['telephone', 'email', 'address_1', 'address_2', 'city', 'country', 'post_code']
+
+    class Meta:
+        abstract = True
+
+
+class LinkFields(models.Model):
+    link_external = models.URLField("External link", blank=True)
+    link_page = models.ForeignKey(
+        'wagtailcore.Page',
+        null=True,
+        blank=True,
+        related_name='+'
+    )
+    link_document = models.ForeignKey(
+        'wagtaildocs.Document',
+        null=True,
+        blank=True,
+        related_name='+'
+    )
+
+    @property
+    def link(self):
+        if self.link_page:
+            return self.link_page.url
+        elif self.link_document:
+            return self.link_document.url
+        else:
+            return self.link_external
+
+    panels = [
+        FieldPanel('link_external'),
+        PageChooserPanel('link_page'),
+        DocumentChooserPanel('link_document'),
+    ]
+
+    api_fields = ['link_external', 'link_page', 'link_document']
+
+    class Meta:
+        abstract = True
+
+
+class RelatedLink(LinkFields):
+    title = models.CharField(max_length=255, help_text="Link title")
+
+    panels = [
+        FieldPanel('title'),
+        MultiFieldPanel(LinkFields.panels, "Link"),
+    ]
+
+    api_fields = ['title'] + LinkFields.api_fields
+
+    class Meta:
+        abstract = True
