+++ blacktail/models/streamfield.py
@@ -1,15 +1,14 @@
 from __future__ import absolute_import, unicode_literals
 
-from django.db import models
 from django import forms
 
-from wagtail.admin.edit_handlers import FieldPanel, MultiFieldPanel, PageChooserPanel
-from wagtail.documents.edit_handlers import DocumentChooserPanel
-
 from wagtail.core.blocks import TextBlock, StructBlock, StreamBlock, FieldBlock, \
-    CharBlock, RichTextBlock, RawHTMLBlock
+    CharBlock, RichTextBlock, RawHTMLBlock, BooleanBlock
 from wagtail.images.blocks import ImageChooserBlock
 from wagtail.documents.blocks import DocumentChooserBlock
+from wagtail.embeds.blocks import EmbedBlock
+
+from .fields import LinkFields, RelatedLink
 
 
 class PullQuoteBlock(StructBlock):
@@ -21,7 +20,7 @@ class PullQuoteBlock(StructBlock):
 
 class ImageFormatChoiceBlock(FieldBlock):
     field = forms.ChoiceField(choices=(
-        ('left', 'Wrap left'), ('right', 'Wrap right'), ('mid', 'Mid width'), ('full', 'Full width'),
+        ('left', 'Wrap left'), ('right', 'Wrap right'), ('mid', 'Mid width'), ('full', 'Full width')
     ))
 
 class HTMLAlignmentChoiceBlock(FieldBlock):
@@ -31,8 +30,32 @@ class HTMLAlignmentChoiceBlock(FieldBlock):
 
 class ImageBlock(StructBlock):
     image = ImageChooserBlock()
-    caption = RichTextBlock()
+    caption = RichTextBlock(required=False)
     alignment = ImageFormatChoiceBlock()
+
+class EmbeddedImageBlock(StructBlock):
+    image = ImageChooserBlock()
+    clickable = BooleanBlock(required=False)
+    caption = CharBlock(required=False)
+
+class HalfImageBlock(StructBlock):
+    image = ImageChooserBlock()
+    h1 = CharBlock(required=False)
+    h2 = CharBlock(required=False)
+    h3 = CharBlock(required=False)
+    text_position = forms.ChoiceField(choices=(
+        ('left', 'Left'), ('right', 'Right')
+    ), required=False)
+
+class FullImageBlock(StructBlock):
+    image = ImageChooserBlock()
+    h1 = CharBlock(required=False)
+    h2 = CharBlock(required=False)
+    h3 = CharBlock(required=False)
+    text = RichTextBlock(required=False)
+    text_position = forms.ChoiceField(choices=(
+        ('left', 'Left'), ('right', 'Right')
+    ), required=False)
 
 class AlignedHTMLBlock(StructBlock):
     html = RawHTMLBlock()
@@ -41,86 +64,18 @@ class AlignedHTMLBlock(StructBlock):
     class Meta:
         icon = "code"
 
-class BlogStreamBlock(StreamBlock):
+
+class StoryStreamBlock(StreamBlock):
+    subheadline = CharBlock(icon="title", classname="title")
+    embedded_image = EmbeddedImageBlock()
+    half_image = HalfImageBlock()
+    full_image = FullImageBlock()
     h2 = CharBlock(icon="title", classname="title")
     h3 = CharBlock(icon="title", classname="title")
     h4 = CharBlock(icon="title", classname="title")
     intro = RichTextBlock(icon="pilcrow")
     paragraph = RichTextBlock(icon="pilcrow")
     aligned_image = ImageBlock(label="Aligned image", icon="image")
-    pullquote = PullQuoteBlock()
     aligned_html = AlignedHTMLBlock(icon="code", label='Raw HTML')
+    embed = EmbedBlock(help_text="URL for media to embed")
     document = DocumentChooserBlock(icon="doc-full-inverse")
-
-class LinkFields(models.Model):
-    link_external = models.URLField("External link", blank=True)
-    link_page = models.ForeignKey(
-        'wagtailcore.Page',
-        null=True,
-        blank=True,
-        related_name='+'
-    )
-    link_document = models.ForeignKey(
-        'wagtaildocs.Document',
-        null=True,
-        blank=True,
-        related_name='+'
-    )
-
-    @property
-    def link(self):
-        if self.link_page:
-            return self.link_page.url
-        elif self.link_document:
-            return self.link_document.url
-        else:
-            return self.link_external
-
-    panels = [
-        FieldPanel('link_external'),
-        PageChooserPanel('link_page'),
-        DocumentChooserPanel('link_document'),
-    ]
-
-    api_fields = ['link_external', 'link_page', 'link_document']
-
-    class Meta:
-        abstract = True
-
-class ContactFields(models.Model):
-    telephone = models.CharField(max_length=20, blank=True)
-    email = models.EmailField(blank=True)
-    address_1 = models.CharField(max_length=255, blank=True)
-    address_2 = models.CharField(max_length=255, blank=True)
-    city = models.CharField(max_length=255, blank=True)
-    country = models.CharField(max_length=255, blank=True)
-    post_code = models.CharField(max_length=10, blank=True)
-
-    panels = [
-        FieldPanel('telephone'),
-        FieldPanel('email'),
-        FieldPanel('address_1'),
-        FieldPanel('address_2'),
-        FieldPanel('city'),
-        FieldPanel('country'),
-        FieldPanel('post_code'),
-    ]
-
-    api_fields = ['telephone', 'email', 'address_1', 'address_2', 'city', 'country', 'post_code']
-
-    class Meta:
-        abstract = True
-
-# Related links
-class RelatedLink(LinkFields):
-    title = models.CharField(max_length=255, help_text="Link title")
-
-    panels = [
-        FieldPanel('title'),
-        MultiFieldPanel(LinkFields.panels, "Link"),
-    ]
-
-    api_fields = ['title'] + LinkFields.api_fields
-
-    class Meta:
-        abstract = True
